Final commit of main.cpp changes
This is the final edits to the project. Main.cpp were edited along with
some minor errors found in some of the weapon classes.

+++ CrazyRandomSword.cpp
@@ -15,7 +15,7 @@ double CrazyRandomSword::hit(double armor) {
 
 	randomHit = rand() % 100 + 7;
 
-	randomArmor = rand() % (0.66 * armor) + 2;
+	randomArmor = rand() % (int) (0.66 * armor) + 2.0;
 
 	double damage = randomHit - randomArmor;
